scripts: make-kernels: import old make-kernels utility for reference
This utility is not used anymore but is easier to use and understand than
its successor build-kernel-list. It's included here more for documentation
purposes than anything else, in case it can help someone by providing some
generic code or help understand the kernel's build process.

Oldest version rescued was v0.5 (2002/04/29), it was used to build 2.4
kernels. The code is covered by the GPL.