scripts: build-kernel-list: import oldest version of build-kernel-list
This is version 0.2 as of 2005-12-05 16:48:48. First version was lost,
we only have this one at the time of forking for version 0.3. This
one is dedicated to building 2.4 kernels. It iterates over a lost of
configuration files and builds and packages the associated kernels.
It's more or less a combination of make-kernels+inst-kernel in a loop
which takes its input arguments from the config file name.