* pkg-0.8.0
Major update : pkg is now able to work with .flxpkg pointing to a directory
which contains a Version file. In this case, this file gets updated for every
"pkg newpkg", and a "pkg release" will simply copy the whole directory to the
global repository. This makes it really easier to support self-contained
source directories for which no source package exists.