* mksp-pkg version 0.0.8
mksp-pkg was too slow when iterating through long lists of packages to
search in large directories, because it would fire a "find" for each of
these packages.

The workaround is now to find only directories starting with the first
word of the package name, and to look inside those ones only. If nothing
is found, the search is performed with the full list as usual. The time
to build an appliance on NFS went down from 15 minutes to less than 1
minute.