scripts: build-kernel-list: don't check +x anymore on add-ons
We got caught several times with svn checkouts which do not restitute
file permissions, causing the add-ons not to be built and the resulting
kernel to silently be incorrect.

Remove the check for -x and use -e instead. That means that it will
not be possible anymore to simply chmod a directory to temporarily
disable the add-on.