scripts: build-kernel-list: add support for multiple platforms
It now becomes possible to specify a platform using -p, which will
cause the .preinit to be fetched from another directory. The name
of the kernel image can be specified using -b (defaults to bzImage),
and the make command can be forced using -m.

The CROSS_COMPILE prefix is used when stripping modules.

Note that this version was in the 2.6-scripts directory and differed
from the one in the git repository in that the SUDO variable was not
called _SUDO, and _DEPMOD did not exist. The merge looks OK though.

The version was bumped to 0.0.10