pkg: fix gcc testing for the prepack command
some software such as git check for CFLAGS changes during a make install,
causing prepack to rebuild everything if it uses different CFLAGS than
compile. This typically happens due to TESTGCC being set only in compile
and not in PREPACK. So make prepack be part of the tested commands to
address this issue.