pkg: correctly use -mcpu instead of -march+-mtune on arm
On ARM, either we use -mcpu to support a specific CPU with its
instruction set and extensions, or we use -march for a generic
family optionally followed by -mtune to optimize for one CPU in
this family.

So when we have CPU set, better use -mcpu to optimize for this CPU
and support its extensions (eg: integer divide, etc...), and fall
back to arch when CPU is not known.