scripts: move all obsolete scripts to attic/
A number of scripts were still present while they're not used anymore
with the firmware packaging mode. These are the scripts used till 2004
used to build a CD or CompactFlash image from a rootfs, and all the
surrounding tools needed to take care of the PRF updates, burn a CD,
or save kernel modules on the rootfs.

Additionally, all scripts dedicated to kernel 2.4 aren't used anymore
either.

Move all of these into the attic directory. Some of them contain useful
functions or can be used as references to debug the other ones, but they
are very unlikely to be reused as is.