scripts: add the pkg-build script
This script builds either one package and its dependencies, or
everything. It's mostly meant to be called from build_firmware, but
may also be called by hand.

Currently it relies on a huge list of variables which all need to be
exported from the caller. It would be wise to change this and to enforce
the local directory structure a little bit.