pkg: support parsing the bld file on-the-fly
pkg only uses $PKGSEARCHDEP during "pkg prepare". The PKGSEARCHDEP directory
is in fact a collection of directories to lookup packages from, but it is
artificially forced to the filtered directory by the project's environment.

In practice, each time a change is needed, it's necessary to modify the bld
file, and re-rerun pkg-filter, which is not always very convenient during
development.

With this change, it becomes possible to specify the path to a bld file in
variable PKGBLDFILTER, whether it's the project's or another one, and only
this file will be considered for resolving package versions during prepare.

That way it becomes possible to develop using different libs that have not
yet been ported to the whole project without being blocked for complete
builds.