scripts: build-kernel-list: workaround new depmod with old kernels
New depmod from the kmod packages try to create modules.builtin.bin from
modules.builtin, but older kernels do not create this file. This causes
an empty modules.builtin.bin file lying on the file system, which makes
modprobe complain about a fatal error on EOF for each module.

Here we detect the situation by checking for this empty file and removing
it when this happens.