scripts: build-kernel-list: automatically try to fix automatic kernel version
When CONFIG_LOCALVERSION_AUTO is set, the kernel build process
determines a complex naming based on the tree state that does not
match the version determined by the script. This prevents modules
from being properly packaged. This change detects this situation
and tries to infer the kernel version from the directory name in
order to be able to package modules and complete the process.

This only works if there is exactly one directory in the install
directory, which normally is the case.