pkg-build: make a distinction between "devel" and "build" packages
Some packages are only provided to resolve build dependencies while others
are needed to build the file system. We now offer the option to prefix the
architecture list in the LPK with :

  - "devel:" for packages that do not need to be built by default
  - "build:" for those which are not to be used for dependencies

This will allow to have some packages built with different options and/or
architectures depending how they are intended to be used. Currently the
support for this is very minimal, we simply trim "build:" and skip
packages tagged "devel:". But that's already enough to make it possible
to list all packages in the LPK file.