pkg-filter: improve duplicate package detection
If a package is referenced twice in formilux.bld, it causes a double
symlink, one in filtered/ and another one in filtered/$package/ which
is wrong. Change the way the link is created to ensure an error is
triggered when this happens, and suggest the possible cause when the
error is detected.