pkg-build: support restriction and exclusion of packages by models
It's often required to restrict a package to a list of models, or to
disable it for a list of model. It is now possible by setting a 3rd
argument in the LPK files indicating a comma-delimited series of
model names which are the only ones which will build this package.
If instead the package needs to be disabled for a known list of models,
just prefix the list with an exclamation mark.