pkg: remove inappropriate check of existence of the short symlinks (.dep, .lst, .tgz)
This one was used only in do_release in case there is no repository (hence
no check on .tested), so that we don't release a package with no tgz. But
the release is not only a matter of binaries but mostly of build.cfg as
well, so we must not prevent from releasing in this case. Also, this entry
was used only as a canary to know that pkg pack did do something, and since
they were not removed before building, they were meaningless!

Note that this was the last user of these old symlinks.