pkg: automatically split .flxrec for recipes when possible
When pkg detects that a package cache is in use through FLX_PKG_CACHE_DIR
and that the current package is compatible with it, it splits the directories
so that we have two links :
  - .flxrec points to the recipes
  - .flxpkg points to the binary package

In order to ease transition, the binary package still contains links to the
recipes directory. It carefully avoids to link the compiled directory in case
it would still exist. Probably that it would be even safer that it would
clean it when it's encountered.