pkg-build: rebuild package if the recipes have changed
In case of package numbering conflict it may happen that a pkg-cache update
retrieves a more recent package from the central repository than the one the
user is currently developing on. This simple change of version should cause
the package to be rebuilt to ensure that there is no inconsistency between
the locally cached binary package and the officially released one as found
in the repository.

This also helps rebuilding components when the build.cfg changes, eventhough
it cannot detect source file changes of course.