scripts/pkg: use absolute locations when building symlinks to REC dir
When making symlinks from the PKG dir to the REC dir, we must resolve
the symlinks, otherwise it happens that they're done to temporary
directories where the models are currently being built, and disappear
when the build dir is destroyed, which is clearly not the purpose!

The same was applied to pkg-build.