scripts/pkg-filter: add support for including the source files sigs in versions
We create a new tag "SRCSIG" which can appear in LPK files. It will be
automatically replaced with the signature of the original component's
source files and recipes. The idea here is that since local components
are not versionned, it's always very hard to detect changes and very
commonly the developer forgets to delete the binaries before rebuilding
and wastes time testing images that do not include local changes.

Since components are generally very small and on the local file system,
it's relatively easy to sign all of their source files, the recipes and
the relevant file names and links and produce a unique signature which
represents the component's state at least for the local machine. That
way, almost all relevant changes can be detected and cause a component
to be rebuilt.