scripts/pkg: explicitly retry source extraction using XZ format after a failure
Some recent versions of TAR (1.23) fail to properly detect the XZ format an
cannot extract .src when the format is .xz (eg: coreutils-8.23). Note that
we cannot use a file signature either since tools like "file" output
different results and formats depending on distros and/or versions. In order
to fix this, we then have to explicitly retry extraction in XZ format after
a first failed attempt.

It was verified to work on a Red Hat 6 where it previously used to fail.