scripts/pkg: fix file type detection on scripts
pkg uses the file command to detect the type of a script, but it
explicitly looks for " script " in the output of the file command,
while newer versions output "Foo script, ASCII ..." for example,
so many scripts are not detected as such anymore.

Fix this by checking for the comma as well. We also run get_perl_depend
when a perl script is detected without bin/env in the sheband.