scripts/pkg: prevent wget from continuing a failed download
"pkg source" used to call "wget -c" which is a very bad idea because if
a download has failed, we cannot trust the end of the file and it becomes
hard to fix the issue. Let's remove -c and in order to avoid jokes with
wgetrc, remove the destination first.