scripts: kernel: provide updated pre-built preinits for multiple architectures
The previous prebuilt init was limited because it required busybox to
be built and packaged for each architecture and that resulted in a big
executable that didn't make it easy to coexist with an external initrd
providing busybox for debugging.

Now that flxutils' preinit implements tar, there's no need anymore for
busybox and we can easily provide all pre-built init codes for several
common architectures. As a matter of compatibility, the "inst-mod26"
directory continues to provide i386 support, "inst-mod26-x86_64" now
provides the x86_64 version (for kernels built without support for 32
bits), "inst-mod26-armv5" provides support for armv5 in EABI mode,
"inst-mod26-armv7" provides support for armv7 in Thumb2 mode, and
"inst-mod26-mips24c" provides support for MIPS24Kc in big endian mode.