script: kernel: support passing an absolute path for the platform
Sometimes it's desirable to use a different preinit coden and it's quite
complicated to modify the final initramfs. Instead it's simpler to support
specifying an absolute path for the platform. If the platform starts with
a "/", it will designate a directory from which .preinit, init, and
possibly busybox will be copied.