script: kernel: build init on the fly if only sources are present
Now that we don't need a specific libc anymore to build init, we
can afford to build it on the fly while packaging the kernel since
we already have the compiler available. So if the init executable
is not found and we find init.c and nolibc.h, then we try to build
it this way. This makes it much easier to deal with various
platforms as there's no need anymore to store the pre-build init
in different subdirectories.