scripts: build-kernel-list: add support for 2-component versions
Starting with 3.x, kernels are numberred with 2 components, the 3rd one
being used for stable versions. Till now we were forced to append .0, but
this change makes the "sub" part optional.

The version is also computed before possibly replacing the config file
name and not after, so that we don't depend on fake config names to
compute this version.