scripts: pkg release: support ssh rsync urls
"pkg release" use to work with a local path as destination for repo (i.e.
worked with nfs mount point).
The patch allows to set repopath as an scp path (i.e. host:/path). If
so, rsync is done from local to remote host, touch and tests on filepath are run on the remote host (has
given in the path).
If repopath is a local path (i.e /path), the behaviour is the same as
before.