scripts: pkg: support schemed repo path
pkg and pkg-cache now support defining a repo by its ssh or file url such as:
	ssh://<server>/<path>
	file://<path>

Note that these formats still work:
	<path>
	<sshserver>:<path>

Status code 3 is used in case the URL cannot be parsed. Note that
the file:// representation requires a path starting with a '/', hence
results in 3 consecutive '/' (like in browsers).