scripts: pkg: only present choices when connected to a terminal
Some scripts fail under pkg because by default when a package is not
found, a list is proposed and a prompt is presented. We should only
do this when both stdin and stdout are a terminal (i.e. interactive
mode) and not a pipe or a file. This will avoid stupidities met in
the past such as "grep foobar $(pkg home blah)" returning whatever.

This patch changes the default PKG_NOCHOICE variable to be zero only
when both in/out are a terminal.