scripts/pkg: add "pkg prefetch" to prefetch a package's source
This can be used to check availability and/or to preload into the cache.
Cacheable methods will be available for later. A good way to use it is
by doing this to prefetch all sources before starting to build:

   pkg-filter list | xargs -n 1 pkg prefetch

Or even better, with error reporting:

  pkg-filter list | grep -v -- -0local- | while read; do
    pkg prefetch $REPLY || { echo "## Failed to prefetch '$REPLY'."; break; }
  done

Note that the downloads must not be run concurrently because the ".src"
file is still being used from the current directory to store the temporary
download.