scripts/pkg: make pkg diff automatically use the last release package
When no argument is passed to "pkg diff", it will now check the changelog
for the last released version and use it as a base for the diff. Practically
speaking this means that issuing "pkg diff" on a package currently being
edited will show the changes since the last release, i.e. the uncommitted
changes.

If it cannot figure the previous version it will just print an error
message inviting the user to pass the package name.