scripts: build-kernel-list: support compressing images
On ARM/ARM64, the image is often "Image" and it is not compressed. But
modern u-boot supports gzip, lzo, lzma and zstd, all of which bring nice
savings to the total image size. With this patch we permit the caller
to specify IMGCOMP= with a command to compress the image, and IMGEXT=
with the extension.