scripts/pkg: improve the error message for released packages
When manually copying a package from a remote repository, it's common
to see "pkg pack" fail saying "the package is already released". In
this case the package repository shows a RELEASE file and there are no
other indications about the expectations. Let's indicate that this error
is only relevant for a local package and that if it comes from a remote
repo, a .repository file might also be missing. Indicate how to proceed
to fix this very common situation.