scripts: kernel: update the copies of init and nolibc for new platforms
The old nolibc didn't support riscv while the new one from kernel 6.9-rc6
does. Also the init tool had been used in a slightly more recent version
for a while but was never updated here.

This was validated on a riscv64 platform with gcc-11.4 where modules are
correctly extracted and everything seems to work as expected.