scripts: kernel: always run make prepare to set the version
If a "make something" is called after building a kernel (e.g. "make dtbs"),
the version is reset, and running the next "build-kernel-list -i" to just
rebuild the image (e.g. after changing compression) will not use the custom
version and will install the modules in a directory named after the default
version instead of the custom one.

Let's just always run make prepare to always set the version in order to
avoid this.