scripts: restore support for globbing in flxrescan/flxextract
While originally it was possible to extract all files from a specific
package, this got lost when better checks were added. It was discussed
lately about reintroducing this. This requires to be able to process
multiple entries for a single one. This is what this patch does. Its
only limitation is that at least one file must match, so it's not
possible to opportunistically use "/path/to/*" and support empty dirs.
Also, "**" is supported to match across directories so it can be an
option to support optionally empty directories, e.g. if a component
needs to match dirs and files.