scripts: kernel: update the preinit script
The new script is more versatile. It ignores root= for initrds so as
not to interfer with regular operating systems. When initrds are not
used, then it uses root= and also respects rootfstype (defaults to
"auto" now) and rootdelay. When none of this is available but "img"
is passed, it will try to find an XIP file system and use it. The
busybox-based debugging part is only mentioned if busybox is found.